test suite refactoring
